factored out raw to clean org and fixed unittest cases
all unittest cases now pass (except for ones expected to not pass)
pylint still says cannot import person for some reason (even though the code runs fine)
deleted a testcase in name_preprocessing (@Stephan did this?) - arguments in assertEqual didn't make sense
